script fix
moved script content from the root into the project
