+++ Assets/Scripts/CameraOffset.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7266dc75d2ea38349863da858c783ce4
+timeCreated: 1481652947
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
